requirements.txt
discord.py==2.3.2
openai==0.27.8
requests==2.31.0
beautifulsoup4==4.12.2
newspaper3k==0.2.8
